Changing amountToPay to BigDecimal

+++ backend/rest/MerchantRest/src/main/java/app/services/InsuranceService.java
@@ -1,5 +1,6 @@
 package app.services;
 
+import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -337,7 +338,7 @@ public class InsuranceService {
 
 		}
 
-		insurance.setAmountToPay(amountToPay);
+		insurance.setAmountToPay(BigDecimal.valueOf(amountToPay));
 
 		return insurance;
 	}
